✨ moving evaluation to db_results app
