+++ tracex_project/db_results/static/db_results/js/filter_toggle.js
