+++ tracex_project/db_results/templates/db_results_overview.html
@@ -15,7 +15,7 @@
         <a href="{% url 'metrics_overview' %}">
             <button class="menu_button">Go to Metrics Dashboard</button>
         </a>
-        <a>
+        <a href="{% url 'evaluation' %}">
             <button class="menu_button">Go to Evaluation</button>
         </a>
     </div>
