+++ tracex_project/db_results/templates/evaluation.html
@@ -8,7 +8,7 @@
     <link href="https://cdn.jsdelivr.net/npm/nouislider@14.6.3/distribute/nouislider.min.css" rel="stylesheet">
     <script src="https://cdn.jsdelivr.net/npm/nouislider@14.6.3/distribute/nouislider.min.js"></script>
     <script src="https://code.jquery.com/jquery-3.6.4.min.js"></script>
-    <script src="{% static '/extraction/js/filter_toggle.js' %}" defer></script>
+    <script src="{% static '/db_results/js/filter_toggle.js' %}" defer></script>
     <link rel="stylesheet" type="text/css" href="{% static '/tracex/css/style.css' %}">
     <style>
         h1 {
@@ -123,6 +123,10 @@
 
     <input type="submit" value="Refresh Filter" class="button">
 </form>
+
+<a href="{% url 'db_results_overview' %}">
+        <button class="menu_button">Return to Database Results Overview</button>
+</a>
 
 <a href="{% url 'landing_page' %}">
     <button class="menu_button">Return to TracEX Menu</button>
